requirements.txt
Django==4.1.3
psycopg2-binary==2.9.5
pillow==9.3.0

gunicorn==20.1.0
django-phonenumber-field[phonenumbers]==7.0.0
